# [Inductor] Support fusion of chained reductions even if keepdims=True