--- a/torch/_inductor/ir.py
+++ b/torch/_inductor/ir.py
@@ -1170,8 +1170,7 @@ class Reduction(Loops):
             r == 1 for r in original_ranges
         ), f"Only enabled for numel_hint == 1, found {original_ranges=}"
         reindex = View.dynamic_reshape_indexer(
-            original_reduction_ranges,
-            tuple(new_ranges) + tuple(new_reduction_ranges),
+            original_reduction_ranges, tuple(new_ranges) + tuple(new_reduction_ranges)
         )
 
         def wrapper_fn(merged_index, new_reduction_index):