@lw 2024-04-24
This is a further improvement on top of #111122 and #112297.