@lw 2024-04-25
Fp8 benchmarks before:
```
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096]. Benchmark results: Inductor: 0.060921625544267055ms, Eager: 0.638029411764706ms, LN only Inductor: 0.06983622047244094ms.
```
and after:
```
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096]. Benchmark results: Inductor: 0.06093886462882096ms, Eager: 0.6382867647058824ms, LN only Inductor: 0.06983410852713179ms.
```

I guess it's expected there's no difference (no improvement) because the codepath used by that benchmark is the one that was already optimized before this PR.