@lw 2024-04-26
I added a case to the benchmark for `keepdim=True`.
Results before this PR:
```
Config: float8_dtype=torch.float8_e4m3fn, shape=[4, 2048, 4096], keepdim=False. Benchmark results: Inductor: 0.06084109149277689ms, Eager: 0.6389044117647059ms, LN only Inductor: 0.06983540372670807ms.
Config: float8_dtype=torch.float8_e4m3fn, shape=[4, 2048, 4096], keepdim=True. Benchmark results: Inductor: 0.11234126984126984ms, Eager: 0.6389044117647059ms, LN only Inductor: 0.06985804416403785ms.
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096], keepdim=False. Benchmark results: Inductor: 0.06084604316546762ms, Eager: 0.6388175182481752ms, LN only Inductor: 0.06981279251170047ms.
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096], keepdim=True. Benchmark results: Inductor: 0.11239711934156378ms, Eager: 0.6387372262773723ms, LN only Inductor: 0.06983644859813085ms.
```
Results after this PR:
```
Config: float8_dtype=torch.float8_e4m3fn, shape=[4, 2048, 4096], keepdim=False. Benchmark results: Inductor: 0.060794212218649515ms, Eager: 0.6377058823529411ms, LN only Inductor: 0.06985625ms.
Config: float8_dtype=torch.float8_e4m3fn, shape=[4, 2048, 4096], keepdim=True. Benchmark results: Inductor: 0.0608911465892598ms, Eager: 0.6377664233576642ms, LN only Inductor: 0.06983937007874015ms.
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096], keepdim=False. Benchmark results: Inductor: 0.06082326951399116ms, Eager: 0.6379051094890511ms, LN only Inductor: 0.06982115085536549ms.
Config: float8_dtype=torch.float8_e5m2, shape=[4, 2048, 4096], keepdim=True. Benchmark results: Inductor: 0.06088041853512705ms, Eager: 0.6381249999999999ms, LN only Inductor: 0.06983720930232558ms.
```